moved button location so it stays in place
