+++ src/app/Page-Functionality/Calendar.css
@@ -2,6 +2,7 @@
     padding-top: 10px;
     position: relative;
     display: flex;
+    width: 100%;
     justify-content: center;
     flex-direction: row;
     align-items: flex-start;
@@ -10,10 +11,13 @@
 }
 
 .panel {
-    margin: 5%;
-    padding: 0;
+    margin: 5px;
+    padding: 5px;
     display: flex;
     flex-direction: column;
+    overflow: hidden;
+    justify-content: center;
+    align-items: center;
     gap: 10px;
     border-radius: 10px;
     background-color: #fff; 
@@ -108,20 +112,6 @@
 }
 
 #calendarSvg {
-    flex: 1;
+    flex: 2;
     margin-top: 20px;
-}
-
-.budget-button {
-    display: flex;
-    text-align: center;
-    justify-content: center;
-    bottom: 0;
-    margin-top: 10px;
-    margin-bottom: 10px;
-    background-color: rgb(150,150,150);
-    color: black;
-    margin-left: 280px;
-    border-radius: 10px;
-    padding: 10px 20px;
 }
